spider_ann/requirements.txt
APScheduler==3.9.1
lxml==4.9.0
Pillow==9.2.0
PyExecJS==1.5.1
PyMySQL==1.0.2
pytesseract==0.3.10
requests==2.24.0
